requirements.txt
python-gevent
python-flask
nginx
mongodb
python-pymongo
